backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.8.3
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0